backend/requirements.txt
pixeltable
fastapi
uvicorn[standard]
python-dotenv
python-multipart
anthropic
sentence-transformers
openai
Pillow
duckduckgo-search
transformers
torch
